requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pillow==10.4.0
numpy==1.26.4
google-generativeai
python-multipart
